+++ example/lib/main.dart
@@ -1,5 +1,4 @@
 import 'package:empire/empire.dart';
-
 import 'package:flutter/material.dart';
 import 'package:uuid/uuid.dart';
 
@@ -22,7 +21,7 @@ class MyApp extends StatelessWidget {
         primarySwatch: Colors.blue,
       ),
       home: Empire(
-        ApplicationViewModel(),
+        ApplicationViewModel(EmpireStringProperty('Empire Counter Example')),
         onAppStateChanged: () => const Uuid().v1(),
         child: CounterPage(
           viewModel: CounterViewModel(),