+++ lib/empire.dart
@@ -1,7 +1,7 @@
-///A simple yet powerful state management library for [Flutter](https://flutter.dev/).
+///A simple yet powerful state management library for Flutter.
 library empire;
 
-export 'empire_state.dart';
-export 'empire_view_model.dart';
-export 'empire_widget.dart';
-export 'empire_property.dart';
+export 'src/empire.dart';
+export 'src/empire_property.dart';
+export 'src/empire_view_model.dart';
+export 'src/empire_widget.dart';