+++ lib/src/empire_exceptions.dart
@@ -19,3 +19,15 @@ class EmpirePropertyNullValueException extends EmpireException {
   String toString() =>
       'EmpirePropertyNullValueException: The underlying value for ${propertyName ?? type} is null.\nStack: $stack';
 }
+
+///Thrown when an EmpireProperty value is changed and attempts to update the UI, but has not been added to the
+///empireProps property of the associated EmpireViewModel
+///
+class PropertyNotAssignedToEmpireViewModelException extends EmpireException {
+  PropertyNotAssignedToEmpireViewModelException(
+      super.stack, super.propertyName, super.type);
+
+  @override
+  String toString() =>
+      'PropertyNotAssignedToEmpireViewModelException: The value for ${propertyName ?? type} changed and failed to update the UI. Did you forget to add it to empireProps?.\nStack: $stack';
+}