+++ lib/src/empire_list_property.dart
@@ -5,8 +5,21 @@ part of 'empire_property.dart';
 ///Any change to the internal list will send a [EmpireStateChanged] event by default. This includes
 ///automatically triggering a UI rebuild.
 class EmpireListProperty<T> extends EmpireProperty<List<T>> {
-  EmpireListProperty(super.value, super.viewModel, {super.propertyName}) {
+  EmpireListProperty(super.value, {super.propertyName}) {
     _originalValue = List<T>.from(value);
+  }
+
+  ///Factory constructor for initializing an [EmpireListProperty] to an empty [List].
+  ///
+  ///See [EmpireProperty] for [propertyName] usages.
+  ///
+  ///## Example
+  ///
+  ///```dart
+  ///final marsOneVolunteers = EmpireListProperty<People>.empty();
+  ///```
+  factory EmpireListProperty.empty({String? propertyName}) {
+    return EmpireListProperty(<T>[], propertyName: propertyName);
   }
 
   /// The number of objects in this list.
@@ -50,7 +63,7 @@ class EmpireListProperty<T> extends EmpireProperty<List<T>> {
     _value.add(value);
 
     if (notifyChanges) {
-      _viewModel.notifyChanges([EmpireStateChanged.addedToList(value)]);
+      viewModel.notifyChanges([EmpireStateChanged.addedToList(value)]);
     }
   }
 
@@ -68,7 +81,7 @@ class EmpireListProperty<T> extends EmpireProperty<List<T>> {
     _value.addAll(values);
 
     if (notifyChanges) {
-      _viewModel.notifyChanges([EmpireStateChanged.addedAllToList(values)]);
+      viewModel.notifyChanges([EmpireStateChanged.addedAllToList(values)]);
     }
   }
 
@@ -93,7 +106,7 @@ class EmpireListProperty<T> extends EmpireProperty<List<T>> {
     final wasRemoved = _value.remove(value);
 
     if (notifyChanges && wasRemoved) {
-      _viewModel.notifyChanges([EmpireStateChanged.removedFromList(value)]);
+      viewModel.notifyChanges([EmpireStateChanged.removedFromList(value)]);
     }
 
     return wasRemoved;
@@ -117,7 +130,7 @@ class EmpireListProperty<T> extends EmpireProperty<List<T>> {
     final removedValue = _value.removeAt(index);
 
     if (notifyChanges) {
-      _viewModel
+      viewModel
           .notifyChanges([EmpireStateChanged.removedFromList(removedValue)]);
     }
 
@@ -140,7 +153,7 @@ class EmpireListProperty<T> extends EmpireProperty<List<T>> {
     _value.clear();
 
     if (notifyChanges) {
-      _viewModel.notifyChanges([stateChangedEvent]);
+      viewModel.notifyChanges([stateChangedEvent]);
     }
   }
 
@@ -208,6 +221,25 @@ class EmpireListProperty<T> extends EmpireProperty<List<T>> {
   /// ```
   Iterable<E> map<E>(E Function(T) toElement) {
     return _value.map(toElement);
+  }
+
+  /// Returns the [index]th element.
+  ///
+  /// The [index] must be non-negative and less than [length].
+  /// Index zero represents the first element (so `iterable.elementAt(0)` is
+  /// equivalent to `iterable.first`).
+  ///
+  /// May iterate through the elements in iteration order, ignoring the
+  /// first [index] elements and then returning the next.
+  /// Some iterables may have a more efficient way to find the element.
+  ///
+  /// Example:
+  /// ```dart
+  /// final numbers = <int>[1, 2, 3, 5, 6, 7];
+  /// final elementAt = numbers.elementAt(4); // 6
+  /// ```
+  T elementAt(int index) {
+    return _value.elementAt(index);
   }
 
   /// Invokes [action] on each element of this iterable in iteration order.