+++ lib/src/empire_map_property.dart
@@ -5,8 +5,21 @@ part of 'empire_property.dart';
 ///Any change to the internal map will send a [EmpireStateChanged] event by default. This includes
 ///automatically triggering a UI rebuild.
 class EmpireMapProperty<K, V> extends EmpireProperty<Map<K, V>> {
-  EmpireMapProperty(super.value, super.viewModel, {super.propertyName}) {
+  EmpireMapProperty(super.value, {super.propertyName}) {
     _originalValue = Map<K, V>.from(value);
+  }
+
+  ///Factory constructor for initializing an [EmpireMapProperty] to an empty [Map].
+  ///
+  ///See [EmpireProperty] for [propertyName] usages.
+  ///
+  ///## Example
+  ///
+  ///```dart
+  ///final jsonData = EmpireMapProperty<String, dynamic>.empty();
+  ///```
+  factory EmpireMapProperty.empty({String? propertyName}) {
+    return EmpireMapProperty(<K, V>{}, propertyName: propertyName);
   }
 
   ///The map entries in the map
@@ -52,7 +65,7 @@ class EmpireMapProperty<K, V> extends EmpireProperty<Map<K, V>> {
     _value.addAll(other);
 
     if (notifyChanges) {
-      _viewModel.notifyChanges([
+      viewModel.notifyChanges([
         EmpireStateChanged.addedMapToMap(other, propertyName: propertyName)
       ]);
     }
@@ -66,16 +79,16 @@ class EmpireMapProperty<K, V> extends EmpireProperty<Map<K, V>> {
     _value.addEntries([entry]);
 
     if (notifyChanges) {
-      _viewModel.notifyChanges([
+      viewModel.notifyChanges([
         EmpireStateChanged.addedToMap(entry.key, entry.value,
             propertyName: propertyName)
       ]);
     }
   }
 
-  /// Adds all key/value pairs of [newEntries] to this map.
+  /// Adds all key/value pairs of [entries] to this map.
   ///
-  /// If a key of [newEntries] is already in this map,
+  /// If a key of [entries] is already in this map,
   /// the corresponding value is overwritten.
   ///
   /// The operation is equivalent to doing `this[entry.key] = entry.value`
@@ -96,7 +109,7 @@ class EmpireMapProperty<K, V> extends EmpireProperty<Map<K, V>> {
     _value.addEntries(entries);
 
     if (notifyChanges) {
-      _viewModel.notifyChanges([
+      viewModel.notifyChanges([
         EmpireStateChanged.addedEntriesToMap(entries,
             propertyName: propertyName)
       ]);
@@ -111,7 +124,7 @@ class EmpireMapProperty<K, V> extends EmpireProperty<Map<K, V>> {
     _value.addEntries([MapEntry<K, V>(key, value)]);
 
     if (notifyChanges) {
-      _viewModel.notifyChanges([
+      viewModel.notifyChanges([
         EmpireStateChanged.addedToMap(key, value, propertyName: propertyName)
       ]);
     }
@@ -147,7 +160,7 @@ class EmpireMapProperty<K, V> extends EmpireProperty<Map<K, V>> {
     final updatedValue = _value.update(key, update, ifAbsent: ifAbsent);
 
     if (notifyChanges) {
-      _viewModel.notifyChanges([
+      viewModel.notifyChanges([
         EmpireStateChanged.updateMapEntry(
           key,
           originalValue,
@@ -186,7 +199,7 @@ class EmpireMapProperty<K, V> extends EmpireProperty<Map<K, V>> {
     });
 
     if (notifyChanges) {
-      _viewModel.notifyChanges(stateChangedEvents);
+      viewModel.notifyChanges(stateChangedEvents);
     }
   }
 
@@ -206,7 +219,7 @@ class EmpireMapProperty<K, V> extends EmpireProperty<Map<K, V>> {
     final removedValue = _value.remove(key);
 
     if (notifyChanges) {
-      _viewModel.notifyChanges([
+      viewModel.notifyChanges([
         EmpireStateChanged.removedFromMap(
           key,
           removedValue,
@@ -243,7 +256,7 @@ class EmpireMapProperty<K, V> extends EmpireProperty<Map<K, V>> {
     });
 
     if (notifyChanges) {
-      _viewModel.notifyChanges(stateChangedEvents);
+      viewModel.notifyChanges(stateChangedEvents);
     }
   }
 
@@ -261,7 +274,7 @@ class EmpireMapProperty<K, V> extends EmpireProperty<Map<K, V>> {
     _value.clear();
 
     if (notifyChanges) {
-      _viewModel.notifyChanges([stateChangedEvent]);
+      viewModel.notifyChanges([stateChangedEvent]);
     }
   }