+++ test/empire_widget_test.dart
@@ -1,41 +1,31 @@
-import 'package:empire/empire_property.dart';
-import 'package:empire/empire_widget.dart';
-import 'package:empire/empire_state.dart';
-import 'package:empire/empire_view_model.dart';
+import 'package:empire/empire.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'dart:math' as math;
 
 class _ApplicationViewModel extends EmpireViewModel {
-  late EmpireProperty<bool> changed;
-
-  @override
-  void initProperties() {
-    changed = createProperty(false, propertyName: 'changed');
-  }
+  final changed = EmpireProperty<bool>(false, propertyName: 'changed');
 
   void change() => changed(!changed.value);
+
+  @override
+  Iterable<EmpireProperty> get empireProps => [changed];
 }
 
 class _ApplicationSubViewModel extends EmpireViewModel {
-  late final EmpireStringProperty viewModelName;
+  final viewModelName = EmpireStringProperty('SubViewModel');
+
   @override
-  void initProperties() {
-    viewModelName = createStringProperty('SubViewModel');
-  }
+  Iterable<EmpireProperty> get empireProps => [viewModelName];
 }
 
 class _TestViewModel extends EmpireViewModel {
-  late EmpireProperty<String?> firstName;
-  late EmpireProperty<String?> lastName;
-  late EmpireProperty<int> age;
+  final firstName = EmpireProperty<String?>(null);
+  final lastName = EmpireProperty<String?>(null);
+  final age = EmpireIntProperty(1);
 
   @override
-  void initProperties() {
-    firstName = createNullProperty();
-    lastName = createNullProperty();
-    age = createProperty(1);
-  }
+  Iterable<EmpireProperty> get empireProps => [firstName, lastName, age];
 }
 
 class _MyWidget extends EmpireWidget<_TestViewModel> {
@@ -208,5 +198,35 @@ void main() {
     await tester.pumpWidget(mainWidget);
 
     expect(find.text(subViewModel.viewModelName.value), findsOneWidget);
+  });
+
+  testWidgets(
+      'EmpireWidget Test - Increment EmpireIntProperty - Finds Correct Text Widget After Property Change',
+      (tester) async {
+    const int initialAge = 10;
+    viewModel.age(initialAge);
+    await tester.pumpWidget(mainWidget);
+
+    expect(find.text("$initialAge"), findsOneWidget);
+
+    final int newAge = viewModel.age.increment();
+    await tester.pumpAndSettle();
+
+    expect(find.text("$newAge"), findsOneWidget);
+  });
+
+  testWidgets(
+      'EmpireWidget Test - Decrement EmpireIntProperty - Finds Correct Text Widget After Property Change',
+      (tester) async {
+    const int initialAge = 10;
+    viewModel.age(initialAge);
+    await tester.pumpWidget(mainWidget);
+
+    expect(find.text("$initialAge"), findsOneWidget);
+
+    final int newAge = viewModel.age.decrement();
+    await tester.pumpAndSettle();
+
+    expect(find.text("$newAge"), findsOneWidget);
   });
 }